Embed fonts, rewrite asset handling
