+++ embedded/css/custom.css
@@ -1,0 +1,163 @@
+@font-face {
+    font-family: 'Iosevka Etoile Web';
+    font-display: swap;
+    font-weight: 400;
+    font-stretch: normal;
+    font-style: normal;
+    src: url('../font/iosevka-etoile-regular.woff2') format('woff2'), url('../font/iosevka-etoile-regular.woff') format('woff'), url('../font/iosevka-etoile-regular.ttf') format('truetype');
+}
+
+@font-face {
+    font-family: 'Iosevka Etoile Web';
+    font-display: swap;
+    font-weight: 400;
+    font-stretch: normal;
+    font-style: italic;
+    src: url('../font/iosevka-etoile-italic.woff2') format('woff2'), url('../font/iosevka-etoile-italic.woff') format('woff'), url('../font/iosevka-etoile-italic.ttf') format('truetype');
+}
+
+@font-face {
+    font-family: 'Iosevka Etoile Web';
+    font-display: swap;
+    font-weight: 700;
+    font-stretch: normal;
+    font-style: normal;
+    src: url('../font/iosevka-etoile-bold.woff2') format('woff2'), url('../font/iosevka-etoile-bold.woff') format('woff'), url('../font/iosevka-etoile-bold.ttf') format('truetype');
+}
+
+@font-face {
+    font-family: 'Iosevka Slab Web';
+    font-display: swap;
+    font-weight: 400;
+    font-stretch: expanded;
+    font-style: normal;
+    src: url('../font/iosevka-fixed-slab-extended.woff2') format('woff2'), url('../font/iosevka-fixed-slab-extended.woff') format('woff'), url('../font/iosevka-fixed-slab-extended.ttf') format('truetype');
+}
+
+.markdown-body {
+    margin: 0 auto;
+    max-width: 738px;
+    font-family: "Iosevka Etoile Web", sans-serif;
+}
+
+tt, code, pre {
+    font-family: "Iosevka Slab Web", monospace !important;
+    font-size: 12px !important;
+}
+
+h1 {
+    border-bottom: none !important;
+}
+
+@media only screen and (max-width: 500px)  {
+    article {
+        padding: 10px;
+    }
+}
+
+/*
+Gruvbox style (light) (c) Pavel Pertsev (original style at https://github.com/morhetz/gruvbox)
+*/
+
+.hljs {
+    display: block;
+    overflow-x: auto;
+    padding: 0.5em;
+    background: #fbf1c7 !important;
+}
+
+.hljs,
+.hljs-subst {
+    color: #3c3836;
+}
+
+/* Gruvbox Red */
+.hljs-deletion,
+.hljs-formula,
+.hljs-keyword,
+.hljs-link,
+.hljs-selector-tag {
+    color: #9d0006;
+}
+
+/* Gruvbox Blue */
+.hljs-built_in,
+.hljs-emphasis,
+.hljs-name,
+.hljs-quote,
+.hljs-strong,
+.hljs-title,
+.hljs-variable {
+    color: #076678;
+}
+
+/* Gruvbox Yellow */
+.hljs-attr,
+.hljs-params,
+.hljs-template-tag,
+.hljs-type {
+    color: #b57614;
+}
+
+/* Gruvbox Purple */
+.hljs-builtin-name,
+.hljs-doctag,
+.hljs-literal,
+.hljs-number {
+    color: #8f3f71;
+}
+
+/* Gruvbox Orange */
+.hljs-code,
+.hljs-meta,
+.hljs-regexp,
+.hljs-selector-id,
+.hljs-template-variable {
+    color: #af3a03;
+}
+
+/* Gruvbox Green */
+.hljs-addition,
+.hljs-meta-string,
+.hljs-section,
+.hljs-selector-attr,
+.hljs-selector-class,
+.hljs-string,
+.hljs-symbol {
+    color: #79740e;
+}
+
+/* Gruvbox Aqua */
+.hljs-attribute,
+.hljs-bullet,
+.hljs-class,
+.hljs-function,
+.hljs-function .hljs-keyword,
+.hljs-meta-keyword,
+.hljs-selector-pseudo,
+.hljs-tag {
+    color: #427b58;
+}
+
+/* Gruvbox Gray */
+.hljs-comment {
+    color: #928374;
+}
+
+/* Gruvbox Purple */
+.hljs-link_label,
+.hljs-literal,
+.hljs-number {
+    color: #8f3f71;
+}
+
+.hljs-comment,
+.hljs-emphasis {
+    font-style: italic;
+}
+
+.hljs-section,
+.hljs-strong,
+.hljs-tag {
+    font-weight: bold;
+}
